refactor: ChartDataServiceをvisualizationディレクトリに移動
- ChartDataService.tsとChartDataService.test.tsをapp/src/application/services/visualization/に移動
- データ可視化関連サービスを専用ディレクトリに集約し組織化を改善
- PerformanceAnalysis.tsxのimport文パスを新しい場所に合わせて更新
- 移動後のファイル内で相対パス調整を実施
- リント、ビルド、テスト全てが成功することを確認
