+++ app/src/presentation/components/PerformanceAnalysis.tsx
@@ -3,9 +3,9 @@
  */
 import { useState } from 'react'
 
-import { ChartDataService } from '../../application/services/ChartDataService'
-import type { PerformanceMetricType } from '../../application/services/ChartDataService'
 import type { PerformanceStatistics } from '../../application/services/PerformanceAnalysisService'
+import { ChartDataService } from '../../application/services/visualization/ChartDataService'
+import type { PerformanceMetricType } from '../../application/services/visualization/ChartDataService'
 import type { PerformanceReport } from '../../domain/models/ai/types'
 import { LineChart } from './charts/LineChart'
